invest: Add chart with stock price history (#680)
* invest: Add chart with stock price history

Adds a chart window that shows stock price history over a configurable
time scale.

* invest: Use array index for x-axis

This gives all data points equal spacing and so gaps between market
closures are not shown.

* invest: Clean up chart pointers on applet removal

Running in-process mode, when the applet was removed from the panel, it
would try to call gtk_widget_get_visible on a destroyed widget, causing
a NULL pointer error. We fix it by properly cleaning up pointers on
applet destroy.